pyproject.toml
[project]
name = "a2a-agents"
version = "0.1.0"
description = "Minimal multi-agent A2A starter with an ADK orchestrator."
requires-python = "==3.13.*"
dependencies = [
    "a2a-sdk>=0.3.25",
    "ddgs>=9.11.4",
    "google-adk>=1.27.1",
    "httpx>=0.28.1",
    "langchain>=1.2.12",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.11",
    "litellm>=1.75.5",
    "uvicorn>=0.42.0",
]
